+++ metagraph/plugins/core/algorithms/__init__.py
