+++ metagraph/plugins/core/algorithms/bipartite.py
@@ -1,5 +1,5 @@
 from metagraph import abstract_algorithm
-from metagraph.types import Graph, BipartiteGraph
+from ..types import Graph, BipartiteGraph
 
 
 @abstract_algorithm("bipartite.graph_projection")
